+++ src/main/java/com/jaquadro/minecraft/storagedrawers/item/ItemDrawerPuller.java
@@ -1,8 +1,24 @@
 package com.jaquadro.minecraft.storagedrawers.item;
 
+import com.jaquadro.minecraft.storagedrawers.config.CommonConfig;
+import net.minecraft.ChatFormatting;
+import net.minecraft.network.chat.Component;
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.api.distmarker.OnlyIn;
+import org.jetbrains.annotations.NotNull;
+
 public class ItemDrawerPuller extends ItemKey
 {
     public ItemDrawerPuller (Properties properties) {
         super(properties);
+    }
+
+    @Override
+    @OnlyIn(Dist.CLIENT)
+    @NotNull
+    public Component getDescription() {
+        return CommonConfig.GENERAL.enableDetachedDrawers.get()
+            ? Component.translatable("item.storagedrawers.drawer_puller.desc")
+            : Component.translatable("itemConfig.storagedrawers.disabled_tool").withStyle(ChatFormatting.RED);
     }
 }