+++ src/main/java/com/jaquadro/minecraft/storagedrawers/item/ItemUpgrade.java
@@ -1,11 +1,14 @@
 package com.jaquadro.minecraft.storagedrawers.item;
 
 import net.minecraft.ChatFormatting;
+import net.minecraft.core.BlockPos;
 import net.minecraft.network.chat.Component;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.level.Level;
+import net.minecraft.world.level.LevelReader;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
 import org.jetbrains.annotations.NotNull;
@@ -49,6 +52,11 @@ public class ItemUpgrade extends Item
     @NotNull
     public Component getDescription() {
         return Component.translatable(this.getDescriptionId() + ".desc");
+    }
+
+    @Override
+    public boolean doesSneakBypassUse (ItemStack stack, LevelReader level, BlockPos pos, Player player) {
+        return true;
     }
 
     public void setAllowMultiple (boolean allow) {