+++ src/main/java/com/jaquadro/minecraft/storagedrawers/item/ItemUpgradeBalance.java
@@ -1,16 +1,15 @@
 package com.jaquadro.minecraft.storagedrawers.item;
 
 import com.jaquadro.minecraft.storagedrawers.config.CommonConfig;
-
 import net.minecraft.ChatFormatting;
 import net.minecraft.network.chat.Component;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
 import org.jetbrains.annotations.NotNull;
 
-public class ItemUpgradePortability extends ItemUpgrade
+public class ItemUpgradeBalance extends ItemUpgrade
 {
-    public ItemUpgradePortability(Properties properties) {
+    public ItemUpgradeBalance(Properties properties) {
         super(properties);
     }
 
@@ -18,8 +17,8 @@ public class ItemUpgradePortability extends ItemUpgrade
     @OnlyIn(Dist.CLIENT)
     @NotNull
     public Component getDescription() {
-        return CommonConfig.GENERAL.heavyDrawers.get()
-            ? Component.translatable("item.storagedrawers.portability_upgrade.desc")
+        return CommonConfig.UPGRADES.enableBalanceUpgrade.get()
+            ? Component.translatable("item.storagedrawers.balance_fill_upgrade.desc")
             : Component.translatable("itemConfig.storagedrawers.disabled_upgrade").withStyle(ChatFormatting.RED);
     }
 }